Move test support classes to TestKit namespace in order to share them with other test cases
