+++ Libraries/CO.CDP.MQ.Tests/Hosting/TestKit/TestServiceScopeFactory.cs
@@ -1,0 +1,11 @@
+using Microsoft.Extensions.DependencyInjection;
+
+namespace CO.CDP.MQ.Tests.Hosting.TestKit;
+
+internal class TestServiceScopeFactory(IServiceProvider services) : IServiceScopeFactory
+{
+    public IServiceScope CreateScope()
+    {
+        return new TestServiceScope(services);
+    }
+}
